refactor(artifacts): isolate platform capability policy
